Add fill and dump commands. Add fill and dump methods on LiquidContainer class.

+++ typeclasses/objects.py
@@ -118,17 +118,35 @@ class Liquid(Consumable):
                  min(PUDDLE_PREFIX.keys(), key=lambda key: abs(key-self.db.uses))]
         self.name = f"{prefix} puddle of {self.db.original_name}"
 
-    def at_drop(self, dropper, **kwargs):
-        super().at_drop(dropper, **kwargs)
-        for obj in dropper.location.contents:
+    # def at_drop(self, dropper, **kwargs):
+    #     super().at_drop(dropper, **kwargs)
+    #     for obj in dropper.location.contents:
+    #         if obj is self:
+    #             continue
+
+    #         if (obj.is_typeclass("typeclasses.objects.Liquid") and 
+    #             obj.db.original_name == self.db.original_name):
+    #             self.db.uses += obj.db.uses
+    #             location = dropper.location
+    #             location.msg_contents(f"{obj.name} is absorbed into {self.name}. {self.name} now has {self.db.uses} units.")
+    #             obj.delete()
+    #     self.set_puddle_name()
+
+    def at_after_move(self, source_location, **kwargs):
+        super().at_after_move(source_location, **kwargs)
+        if not self.location.is_typeclass("typeclasses.rooms.Room"):
+            self.location.msg_contents("Not in a room")
+            return 
+
+        self.location.msg_contents("In a room")
+        for obj in self.location.contents:
             if obj is self:
                 continue
 
-            if (obj.is_typeclass("typeclasses.objects.Liquid") and 
-                obj.db.original_name == self.db.original_name):
+            if (obj.is_typeclass("typeclasses.objects.Liquid")
+                and obj.db.original_name == self.db.original_name):
                 self.db.uses += obj.db.uses
-                location = dropper.location
-                location.msg_contents(f"{obj.name} is absorbed into {self.name}. {self.name} now has {self.db.uses} units.")
+                target_location.msg_contents(f"{obj.name} is absorbed into {self.name}. {self.name} now has {self.db.uses} units.")
                 obj.delete()
         self.set_puddle_name()
 
@@ -144,6 +162,78 @@ class LiquidContainer(ContainerMassMixin, Object):
         self.db.mass_reduction = 1
         self.db.category = "container"
 
+    # def return_contents(self):
+    #     if not self.contents:
+    #         return False
+    #     return {"name":self.contents[0].db.original_name, 
+    #             "units":self.contents[0].db.uses}        
+
+    def fill(self, source, caller=None):
+        # If source is a liquid container, get liquid from container
+        if source.is_typeclass("typeclasses.objects.LiquidContainer"):
+            if not source.contents:
+                if caller:
+                    caller.msg(f"{source.name} has no liquid.")
+                return
+            source = source.contents[0]
+
+        # Check capacity
+        capacity = self.return_capacity()
+        
+        # container has liquid
+        if self.contents:
+            # Is the source liquid the same as the container contents
+            if self.contents[0].db.original_name != source.db.original_name:
+                if caller: caller.msg("Can't fill container with a different liquid")
+                return False
+
+            self.contents[0].db.uses += capacity
+            source.db.uses -= capacity
+        # no liquid in container
+        else:
+            if source.db.uses > capacity:
+                copy = source.copy(source.db.original_name)
+                copy.db.uses = capacity
+                copy.move_to(self, quiet=True)
+                source.db.uses -= capacity
+            else:
+                source.name = source.db.original_name
+                source.move_to(self, quiet=True)
+
+        if caller:
+            caller.msg(f"You fill |w{self.name}|n with |w{source.db.original_name}|n.")
+            rooms.dark_aware_msg(
+                "|w{character}|n fills |w{object}|n with |w{liquid}|n.",
+                caller.location,
+                {"{character}":caller.name, "{object}":self.name, "{liquid}":source.db.original_name},
+                {"{character}":"Someone", "{object}":"something", "{liquid}":"something"},
+                caller
+            )
+
+        if source.db.uses <= 0:
+            source.delete()
+
+        return True
+
+    def dump(self, location, caller=None):
+        if not self.contents:
+            if caller: caller.msg(f"{self.name} has no liquid to dump.")
+            return False
+        
+        self.contents[0].move_to(location, quiet=True)
+        return True
+
+    def return_capacity(self):
+        if not self.contents:
+            return self.db.capacity
+        capacity = self.db.capacity - self.contents[0].db.uses
+        return capacity
+    
+    def return_contents(self):
+        if not self.contents:
+            return {"name": "Nothing", "uses":0}
+        return {"name":self.contents[0].db.original_name, "uses":self.contents[0].db.uses}
+
     def at_object_receive(self, moved_obj, source_location, **kwargs):
         super().at_object_receive(moved_obj, source_location, **kwargs)
 
@@ -152,5 +242,5 @@ class LiquidContainer(ContainerMassMixin, Object):
 
     def return_appearance(self, looker, **kwargs):
         string = super().return_appearance(looker, **kwargs)
-        string += f"\nIt is filled with (liquid). It is half full."
+        string += f"\nIt is filled with {self.return_contents()['name']}. {self.return_contents()['uses']} / {self.db.capacity}"
         return string
